+++ resources/js/components/Fields/FileField.vue
@@ -65,7 +65,7 @@ function getName() {
 </script>
 
 <template>
-    <BaseField v-bind="{ label, name, error, className, help }">
+    <BaseField v-bind="{ label, name, error, className, help, theme }">
       <div class="flex w-full">
           <span
               v-if="prepend"