+++ resources/js/components/Fields/TextField.vue
@@ -55,7 +55,7 @@ function leftRoundClass() {
 </script>
 
 <template>
-  <BaseField v-bind="{ label, name, error, className, help }">
+  <BaseField v-bind="{ label, name, error, className, help, theme }">
     <div class="flex w-full">
       <span
         v-if="prepend"